Move project checks to "project-checker" and fix linter configuration
