+++ package.json
@@ -10,10 +10,10 @@
     "lerna": "^3.20.2",
     "eslint": "^6.8.0",
     "prettier": "^1.19.1",
-    "eslint-config-prettier": "^6.9.0",
+    "eslint-config-prettier": "^6.10.0",
     "eslint-plugin-prettier": "^3.1.2",
-    "@typescript-eslint/eslint-plugin": "^2.17.0",
-    "@typescript-eslint/parser": "^1.12.0",
+    "@typescript-eslint/eslint-plugin": "^2.18.0",
+    "@typescript-eslint/parser": "^2.18.0",
     "typescript": "^3.7.5",
     "chai": "^4.2.0",
     "@types/chai": "^4.2.7",
