+++ packages/cli/src/commands/new/event.ts
@@ -9,9 +9,10 @@ import {
   parseFields,
   ImportDeclaration,
 } from '../../services/generator/target'
-import { generate, checkItIsABoosterProject } from '../../services/generator'
+import { generate } from '../../services/generator'
 import * as path from 'path'
 import { templates } from '../../templates'
+import { checkItIsABoosterProject } from "../../services/project-checker";
 
 export default class Event extends Oclif.Command {
   public static description = 'create a new event'
