+++ packages/cli/tsconfig.json
@@ -5,6 +5,7 @@
     "rootDir": "./src"
   },
   "include": [
-    "src/**/*"
+    "src/**/*",
+    "test/**/*"
   ]
 }
