--- packages/framework-types/test/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../tsconfig",
-  "compilerOptions": {
-    "noEmit": true,
-    "experimentalDecorators": true
-  },
-  "references": [{ "path": ".." }]
-}
